Re-organise utils
